+++ web/src/test/java/com/graphhopper/application/resources/RouteResourceLeipzigTest.java
@@ -23,6 +23,7 @@ import com.graphhopper.application.GraphHopperApplication;
 import com.graphhopper.application.GraphHopperServerConfiguration;
 import com.graphhopper.application.util.GraphHopperServerTestConfiguration;
 import com.graphhopper.config.CHProfile;
+import com.graphhopper.config.Profile;
 import com.graphhopper.routing.TestProfiles;
 import com.graphhopper.util.BodyAndStatus;
 import com.graphhopper.util.Helper;
@@ -42,6 +43,8 @@ import java.util.Random;
 
 import static com.graphhopper.application.resources.Util.getWithStatus;
 import static com.graphhopper.application.util.TestUtils.clientTarget;
+import static com.graphhopper.json.Statement.If;
+import static com.graphhopper.json.Statement.Op.MULTIPLY;
 import static com.graphhopper.util.Parameters.Algorithms.*;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertTrue;
@@ -80,12 +83,12 @@ public class RouteResourceLeipzigTest {
 
     @ParameterizedTest
     @CsvSource(value = {
-            "103,-1,algorithm=" + DIJKSTRA_BI,
-            "128,-1,algorithm=" + ASTAR_BI,
-            "30867,1,ch.disable=true&algorithm=" + DIJKSTRA,
-            "21181,1,ch.disable=true&algorithm=" + ASTAR,
-            "14854,1,ch.disable=true&algorithm=" + DIJKSTRA_BI,
-            "10538,1,ch.disable=true&algorithm=" + ASTAR_BI
+            "90,-1,algorithm=" + DIJKSTRA_BI,
+            "104,-1,algorithm=" + ASTAR_BI,
+            "33133,1,ch.disable=true&algorithm=" + DIJKSTRA,
+            "22497,1,ch.disable=true&algorithm=" + ASTAR,
+            "15434,1,ch.disable=true&algorithm=" + DIJKSTRA_BI,
+            "10926,1,ch.disable=true&algorithm=" + ASTAR_BI
     })
     void testTimeout(int expectedVisitedNodes, int timeout, String args) {
         {